♻️ Refactor imports to streamline file handling and security functions
